+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 This is a history of changes to clara-rules.
 
 ### 0.17.0-SNAPSHOT
+* Add clear-ns-productions! functionality to support clean reloading of rule and query definitions. See [issue 316](https://github.com/cerner/clara-rules/issues/316) for details.
 * Support session inspection and fact-graph in ClojureScript. See [issue 307](https://github.com/cerner/clara-rules/issues/307) for details.
 
 ### 0.16.1